+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Add basic `main`/`app` scaffolding and main loop
 - Add logger
 - Shave off unnecessary dependencies and/or their features
-- Add `anyhow` crate
+- Replace `color-eyre` with`anyhow`
 
 // 602K (615.592)
 
@@ -22,3 +22,12 @@
 - Refactor `main` and `logger`
 
 // 611K (625.328)
+
+### 0.0.3 / 2026-04-05
+
+- First draft of layout
+- Add multicolored in-console logging, lines are scrollable using keyboard and mouse, toggleable via `[features]`
+- Put in-file and in-console logging in a separate module
+- Improve frame rate and responsiveness
+
+// 622K (636.384)